requirements.txt
yt-dlp==2025.1.8
google-api-python-client==2.149.0
openai>=1.40.0
PyYAML>=6.0.1
